requirements.txt
app==0.0.1
attr==0.3.2
colorama==0.4.6
ConfigParser==7.2.0
cryptography==45.0.6
docutils==0.22
faiss==1.5.3
fastapi==0.116.1
filelock==3.19.1
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.7
jnius==1.1.0
keyring==25.6.0
numpy==2.3.2
pdfplumber==0.11.7
Pillow==11.3.0
protobuf==6.32.0
pydantic==2.11.7
pyOpenSSL==25.1.0
pytesseract==0.3.13
redis==6.4.0
sentence_transformers==5.1.0
Sphinx==8.2.3
thread==2.0.5
typer==0.16.1
urllib3_secure_extra==0.1.0
watchdog==6.0.0
xmlrpclib==1.0.1
